Fix Travis CI config
This package needs a Redis server to be tested and, unfortunately,
Travis CI doesn't support services in OSX, making this package testable
only in Linux.

Also, by the time of this commit, Go 1.11 has not been released yet, but
there are beta versions available. It seems go1.11beta3 has weird,
broken behaviors, making go1.11beta3 break the tests, so it is now
removed.